# students/k3339/Gneushev_Vladislav/lab3/docs/swagger.json
{
   "openapi":"3.1.0",
   "info":{
      "title":"FastAPI",
      "version":"0.1.0"
   },
   "paths":{
      "/api/drivers":{
         "get":{
            "tags":[
               "Водители"
            ],
            "summary":"Получить список водителей",
            "operationId":"get_drivers_api_drivers_get",
            "responses":{
               "200":{
                  "description":"Successful Response",
                  "content":{
                     "application/json":{
                        "schema":{
                           "type":"array",
                           "items":{
                              "$ref":"#/components/schemas/GetDriverSchema"
                           },
                           "title":"Response Get Drivers Api Drivers Get"
                        }
                     }
                  }
               }
            }
         },
         "post":{
            "tags":[
               "Водители"
            ],
            "summary":"Добавить водителя",
            "operationId":"add_driver_api_drivers_post",
            "parameters":[
               {
                  "name":"Authorization",
                  "in":"header",
                  "required":true,
                  "schema":{
                     "type":"string",
                     "title":"Authorization"
                  }
               }
            ],
            "requestBody":{
               "required":true,
               "content":{
                  "application/json":{
                     "schema":{
                        "$ref":"#/components/schemas/AddDriverSchema"
                     }
                  }
               }
            },
            "responses":{
               "201":{
                  "description":"Successful Response",
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/GetDriverSchema"
                        }
                     }
                  }
               },
               "422":{
                  "description":"Validation Error",
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         }
      },
      "/api/drivers/{driver_id}":{
         "delete":{
            "tags":[
               "Водители"
            ],
            "summary":"Удалить водителя",
            "operationId":"delete_driver_api_drivers__driver_id__delete",
            "parameters":[
               {
                  "name":"driver_id",
                  "in":"path",
                  "required":true,
                  "schema":{
                     "type":"integer",
                     "title":"Driver Id"
                  }
               },
               {
                  "name":"Authorization",
                  "in":"header",
                  "required":true,
                  "schema":{
                     "type":"string",
                     "title":"Authorization"
                  }
               }
            ],
            "responses":{
               "204":{
                  "description":"Successful Response"
               },
               "422":{
                  "description":"Validation Error",
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         }
      },
      "/api/drivers/classes":{
         "get":{
            "tags":[
               "Водители"
            ],
            "summary":"Получить список классов водителей",
            "operationId":"get_driver_classes_api_drivers_classes_get",
            "responses":{
               "200":{
                  "description":"Successful Response",
                  "content":{
                     "application/json":{
                        "schema":{
                           "type":"array",
                           "items":{
                              "$ref":"#/components/schemas/GetDriverClassSchema"
                           },
                           "title":"Response Get Driver Classes Api Drivers Classes Get"
                        }
                     }
                  }
               }
            }
         },
         "post":{
            "tags":[
               "Водители"
            ],
            "summary":"Добавить класс водителя",
            "operationId":"add_driver_class_api_drivers_classes_post",
            "parameters":[
               {
                  "name":"Authorization",
                  "in":"header",
                  "required":true,
                  "schema":{
                     "type":"string",
                     "title":"Authorization"
                  }
               }
            ],
            "requestBody":{
               "required":true,
               "content":{
                  "application/json":{
                     "schema":{
                        "$ref":"#/components/schemas/AddDriverClassSchema"
                     }
                  }
               }
            },
            "responses":{
               "201":{
                  "description":"Successful Response",
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/GetDriverClassSchema"
                        }
                     }
                  }
               },
               "422":{
                  "description":"Validation Error",
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         }
      },
      "/api/drivers/classes/{driver_class_id}":{
         "delete":{
            "tags":[
               "Водители"
            ],
            "summary":"Удалить класс водителя",
            "operationId":"delete_driver_class_api_drivers_classes__driver_class_id__delete",
            "parameters":[
               {
                  "name":"driver_class_id",
                  "in":"path",
                  "required":true,
                  "schema":{
                     "type":"integer",
                     "title":"Driver Class Id"
                  }
               },
               {
                  "name":"Authorization",
                  "in":"header",
                  "required":true,
                  "schema":{
                     "type":"string",
                     "title":"Authorization"
                  }
               }
            ],
            "responses":{
               "204":{
                  "description":"Successful Response"
               },
               "422":{
                  "description":"Validation Error",
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         }
      },
      "/api/drivers/{driver_id}/work_schedule":{
         "get":{
            "tags":[
               "Водители"
            ],
            "summary":"Получить график работы водителя",
            "operationId":"get_driver_work_schedule_api_drivers__driver_id__work_schedule_get",
            "parameters":[
               {
                  "name":"driver_id",
                  "in":"path",
                  "required":true,
                  "schema":{
                     "type":"integer",
                     "title":"Driver Id"
                  }
               }
            ],
            "responses":{
               "200":{
                  "description":"Successful Response",
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/GetWorkScheduleSchema"
                        }
                     }
                  }
               },
               "422":{
                  "description":"Validation Error",
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         }
      },
...
      "/api/routes":{
         "get":{
            "tags":[
               "Маршруты"
            ],
            "summary":"Получить список маршрутов",
            "operationId":"get_routes_api_routes_get",
            "responses":{
               "200":{
                  "description":"Successful Response",
                  "content":{
                     "application/json":{
                        "schema":{
                           "type":"array",
                           "items":{
                              "$ref":"#/components/schemas/GetRouteSchema"
                           },
                           "title":"Response Get Routes Api Routes Get"
                        }
                     }
                  }
               }
            }
         },
         "post":{
            "tags":[
               "Маршруты"
            ],
            "summary":"Добавить маршрут",
            "operationId":"add_route_api_routes_post",
            "parameters":[
               {
                  "name":"Authorization",
                  "in":"header",
                  "required":true,
                  "schema":{
                     "type":"string",
                     "title":"Authorization"
                  }
               }
            ],
            "requestBody":{
               "required":true,
               "content":{
                  "application/json":{
                     "schema":{
                        "$ref":"#/components/schemas/AddRouteSchema"
                     }
                  }
               }
            },
            "responses":{
               "201":{
                  "description":"Successful Response",
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/GetRouteSchema"
                        }
                     }
                  }
               },
               "422":{
                  "description":"Validation Error",
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         }
      },
      "/api/routes/{route_id}":{
         "delete":{
            "tags":[
               "Маршруты"
            ],
            "summary":"Удалить маршрут",
            "operationId":"delete_route_api_routes__route_id__delete",
            "parameters":[
               {
                  "name":"route_id",
                  "in":"path",
                  "required":true,
                  "schema":{
                     "type":"integer",
                     "title":"Route Id"
                  }
               },
               {
                  "name":"Authorization",
                  "in":"header",
                  "required":true,
                  "schema":{
                     "type":"string",
                     "title":"Authorization"
                  }
               }
            ],
            "responses":{
               "204":{
                  "description":"Successful Response"
               },
               "422":{
                  "description":"Validation Error",
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         }
      },
...
   },
   "components":{
      "schemas":{
...
         "AddDriverClassSchema":{
            "properties":{
               "name":{
                  "type":"string",
                  "title":"Name"
               }
            },
            "type":"object",
            "required":[
               "name"
            ],
            "title":"AddDriverClassSchema"
         },
         "AddDriverSchema":{
            "properties":{
               "driver_class_name":{
                  "type":"string",
                  "title":"Driver Class Name"
               },
               "first_name":{
                  "type":"string",
                  "title":"First Name"
               },
               "last_name":{
                  "type":"string",
                  "title":"Last Name"
               },
               "passport_info":{
                  "type":"string",
                  "title":"Passport Info"
               },
               "work_experience_months":{
                  "type":"integer",
                  "title":"Work Experience Months"
               }
            },
            "type":"object",
            "required":[
               "driver_class_name",
               "first_name",
               "last_name",
               "passport_info",
               "work_experience_months"
            ],
            "title":"AddDriverSchema"
         },
         "AddRouteSchema":{
            "properties":{
               "name":{
                  "type":"string",
                  "title":"Name"
               },
               "start_point_name":{
                  "type":"string",
                  "title":"Start Point Name"
               },
               "end_point_name":{
                  "type":"string",
                  "title":"End Point Name"
               },
               "start_time":{
                  "type":"string",
                  "format":"time",
                  "title":"Start Time"
               },
               "end_time":{
                  "type":"string",
                  "format":"time",
                  "title":"End Time"
               },
               "interval_seconds":{
                  "type":"integer",
                  "title":"Interval Seconds"
               },
               "duration_seconds":{
                  "type":"integer",
                  "title":"Duration Seconds"
               },
               "length_km":{
                  "type":"integer",
                  "title":"Length Km"
               }
            },
            "type":"object",
            "required":[
               "name",
               "start_point_name",
               "end_point_name",
               "start_time",
               "end_time",
               "interval_seconds",
               "duration_seconds",
               "length_km"
            ],
            "title":"AddRouteSchema"
         },
...
         "DayWorkingHoursSchema":{
            "properties":{
               "start_time":{
                  "type":"string",
                  "format":"time",
                  "title":"Start Time"
               },
               "end_time":{
                  "type":"string",
                  "format":"time",
                  "title":"End Time"
               }
            },
            "type":"object",
            "required":[
               "start_time",
               "end_time"
            ],
            "title":"DayWorkingHoursSchema"
         },
...
         "GetDriverClassSchema":{
            "properties":{
               "id":{
                  "type":"integer",
                  "title":"Id"
               },
               "name":{
                  "type":"string",
                  "title":"Name"
               }
            },
            "type":"object",
            "required":[
               "id",
               "name"
            ],
            "title":"GetDriverClassSchema"
         },
         "GetDriverSchema":{
            "properties":{
               "id":{
                  "type":"integer",
                  "title":"Id"
               },
               "first_name":{
                  "type":"string",
                  "title":"First Name"
               },
               "last_name":{
                  "type":"string",
                  "title":"Last Name"
               },
               "passport_info":{
                  "type":"string",
                  "title":"Passport Info"
               },
               "work_experience_months":{
                  "type":"integer",
                  "title":"Work Experience Months"
               },
               "driver_class":{
                  "$ref":"#/components/schemas/GetDriverClassSchema"
               },
               "salary_rub":{
                  "anyOf":[
                     {
                        "type":"integer"
                     },
                     {
                        "type":"null"
                     }
                  ],
                  "title":"Salary Rub"
               }
            },
            "type":"object",
            "required":[
               "id",
               "first_name",
               "last_name",
               "passport_info",
               "work_experience_months",
               "driver_class"
            ],
            "title":"GetDriverSchema"
         },
...
         "GetRouteSchema":{
            "properties":{
               "id":{
                  "type":"integer",
                  "title":"Id"
               },
               "name":{
                  "type":"string",
                  "title":"Name"
               },
               "start_point_name":{
                  "type":"string",
                  "title":"Start Point Name"
               },
               "end_point_name":{
                  "type":"string",
                  "title":"End Point Name"
               },
               "start_time":{
                  "type":"string",
                  "format":"time",
                  "title":"Start Time"
               },
               "end_time":{
                  "type":"string",
                  "format":"time",
                  "title":"End Time"
               },
               "interval_seconds":{
                  "type":"integer",
                  "title":"Interval Seconds"
               },
               "duration_seconds":{
                  "type":"integer",
                  "title":"Duration Seconds"
               },
               "length":{
                  "type":"integer",
                  "title":"Length"
               }
            },
            "type":"object",
            "required":[
               "id",
               "name",
               "start_point_name",
               "end_point_name",
               "start_time",
               "end_time",
               "interval_seconds",
               "duration_seconds",
               "length"
            ],
            "title":"GetRouteSchema"
         },
         "GetWorkScheduleSchema":{
            "properties":{
               "monday":{
                  "anyOf":[
                     {
                        "$ref":"#/components/schemas/DayWorkingHoursSchema"
                     },
                     {
                        "type":"null"
                     }
                  ]
               },
               "tuesday":{
                  "anyOf":[
                     {
                        "$ref":"#/components/schemas/DayWorkingHoursSchema"
                     },
                     {
                        "type":"null"
                     }
                  ]
               },
               "wednesday":{
                  "anyOf":[
                     {
                        "$ref":"#/components/schemas/DayWorkingHoursSchema"
                     },
                     {
                        "type":"null"
                     }
                  ]
               },
               "thursday":{
                  "anyOf":[
                     {
                        "$ref":"#/components/schemas/DayWorkingHoursSchema"
                     },
                     {
                        "type":"null"
                     }
                  ]
               },
               "friday":{
                  "anyOf":[
                     {
                        "$ref":"#/components/schemas/DayWorkingHoursSchema"
                     },
                     {
                        "type":"null"
                     }
                  ]
               },
               "saturday":{
                  "anyOf":[
                     {
                        "$ref":"#/components/schemas/DayWorkingHoursSchema"
                     },
                     {
                        "type":"null"
                     }
                  ]
               },
               "sunday":{
                  "anyOf":[
                     {
                        "$ref":"#/components/schemas/DayWorkingHoursSchema"
                     },
                     {
                        "type":"null"
                     }
                  ]
               }
            },
            "type":"object",
            "required":[
               "monday",
               "tuesday",
               "wednesday",
               "thursday",
               "friday",
               "saturday",
               "sunday"
            ],
            "title":"GetWorkScheduleSchema"
         },
         "HTTPValidationError":{
            "properties":{
               "detail":{
                  "items":{
                     "$ref":"#/components/schemas/ValidationError"
                  },
                  "type":"array",
                  "title":"Detail"
               }
            },
            "type":"object",
            "title":"HTTPValidationError"
         },
...
         "ValidationError":{
            "properties":{
               "loc":{
                  "items":{
                     "anyOf":[
                        {
                           "type":"string"
                        },
                        {
                           "type":"integer"
                        }
                     ]
                  },
                  "type":"array",
                  "title":"Location"
               },
               "msg":{
                  "type":"string",
                  "title":"Message"
               },
               "type":{
                  "type":"string",
                  "title":"Error Type"
               }
            },
            "type":"object",
            "required":[
               "loc",
               "msg",
               "type"
            ],
            "title":"ValidationError"
         }
      }
   }
}